eawatersapp/requirements.txt
aldryn-boilerplates==0.8.0
aldryn-forms==5.0.4
amqp==2.6.1
asgiref==3.2.10
autopep8==1.4.4
billiard==3.6.3.0
cachetools==4.1.1
celery==4.4.7
certifi==2019.3.9
chardet==3.0.4
cssselect==1.1.0
cssutils==1.0.2
defusedxml==0.6.0
Django==2.2.14
django-absolute==0.3
django-allauth==0.39.1
django-appconf==1.0.4
django-celery-results==1.2.1
django-classy-tags==1.0.0
django-cms==3.7.1
django-cors-headers==3.4.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-emailit==0.2.4
django-filer==1.7.1
django-filter==2.3.0
django-formtools==2.2
django-ipware==3.0.1
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-mptt==0.11.0
django-polymorphic==2.1.2
django-sekizai==1.1.0
django-sizefield==1.0.0
django-tablib==3.2
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.6.0
djangocms-column==1.11.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
djangorestframework==3.11.0
easy-thumbnails==2.7
hashids==1.3.1
html5lib==1.1
idna==2.8
jsonfield==3.1.0
kombu==4.6.11
lxml==4.5.2
Markdown==3.2.2
mysqlclient==2.0.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.2.2
pkg-resources==0.0.0
premailer==3.7.0
pycodestyle==2.5.0
PyPDF2==1.26.0
python-decouple==3.1
python-http-client==3.3.1
python3-openid==3.1.0
pytz==2018.5
redis==3.5.3
reportlab==3.5.46
requests==2.21.0
requests-oauthlib==1.2.0
sendgrid==6.4.6
six==1.15.0
sqlparse==0.2.4
starkbank-ecdsa==1.0.0
stripe==2.27.0
tablib==2.0.0
Unidecode==1.1.1
urllib3==1.24.2
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.4
YURL==1.0.0
